+++ src/main/java/blusunrize/immersiveengineering/common/util/compat/crafttweaker/managers/CrusherRecipeManager.java
@@ -10,6 +10,7 @@ package blusunrize.immersiveengineering.common.util.compat.crafttweaker.managers
 import blusunrize.immersiveengineering.api.crafting.CrusherRecipe;
 import blusunrize.immersiveengineering.api.crafting.IESerializableRecipe;
 import blusunrize.immersiveengineering.api.crafting.StackWithChance;
+import blusunrize.immersiveengineering.common.config.IEServerConfig;
 import blusunrize.immersiveengineering.common.util.compat.crafttweaker.CrTIngredientUtil;
 import blusunrize.immersiveengineering.common.util.compat.crafttweaker.actions.AbstractActionRemoveMultipleOutputs;
 import com.blamejared.crafttweaker.api.CraftTweakerAPI;
@@ -105,7 +106,9 @@ public class CrusherRecipeManager implements IRecipeManager<CrusherRecipe>
 
 		final ItemStack result = mainOutput.getInternal();
 		final Ingredient ingredient = input.asVanillaIngredient();
-		final CrusherRecipe recipe = new CrusherRecipe(resourceLocation, of(result), ingredient, energy);
+		final CrusherRecipe recipe = IEServerConfig.MACHINES.crusherConfig.apply(
+				new CrusherRecipe(resourceLocation, of(result), ingredient, energy)
+		);
 
 		for(Percentaged<IItemStack> additionalOutput : additionalOutputs)
 		{