+++ src/main/java/blusunrize/immersiveengineering/common/util/compat/crafttweaker/managers/MetalPressRecipeManager.java
@@ -10,6 +10,7 @@ package blusunrize.immersiveengineering.common.util.compat.crafttweaker.managers
 import blusunrize.immersiveengineering.api.crafting.IESerializableRecipe;
 import blusunrize.immersiveengineering.api.crafting.IngredientWithSize;
 import blusunrize.immersiveengineering.api.crafting.MetalPressRecipe;
+import blusunrize.immersiveengineering.common.config.IEServerConfig;
 import blusunrize.immersiveengineering.common.util.compat.crafttweaker.CrTIngredientUtil;
 import com.blamejared.crafttweaker.api.CraftTweakerAPI;
 import com.blamejared.crafttweaker.api.action.recipe.ActionAddRecipe;
@@ -63,8 +64,8 @@ public class MetalPressRecipeManager implements IRecipeManager<MetalPressRecipe>
 		final IngredientWithSize ingredient = CrTIngredientUtil.getIngredientWithSize(input);
 		final ItemStack outputStack = output.getInternal();
 
-		final MetalPressRecipe recipe = new MetalPressRecipe(
-				resourceLocation, IESerializableRecipe.of(outputStack), ingredient, mold.getDefinition(), energy
+		final MetalPressRecipe recipe = IEServerConfig.MACHINES.metalPressConfig.apply(
+				new MetalPressRecipe(resourceLocation, IESerializableRecipe.of(outputStack), ingredient, mold.getDefinition(), energy)
 		);
 		CraftTweakerAPI.apply(new ActionAddRecipe<>(this, recipe, null));
 	}