+++ src/main/java/blusunrize/immersiveengineering/common/util/compat/crafttweaker/managers/RefineryRecipeManager.java
@@ -9,6 +9,7 @@ package blusunrize.immersiveengineering.common.util.compat.crafttweaker.managers
 
 import blusunrize.immersiveengineering.api.crafting.FluidTagInput;
 import blusunrize.immersiveengineering.api.crafting.RefineryRecipe;
+import blusunrize.immersiveengineering.common.config.IEServerConfig;
 import blusunrize.immersiveengineering.common.util.compat.crafttweaker.CrTIngredientUtil;
 import blusunrize.immersiveengineering.common.util.compat.crafttweaker.actions.AbstractActionGenericRemoveRecipe;
 import blusunrize.immersiveengineering.common.util.compat.crafttweaker.actions.ActionAddRecipeCustomOutput;
@@ -119,6 +120,7 @@ public class RefineryRecipeManager implements IRecipeManager<RefineryRecipe>
 		final FluidTagInput tagInput2 = CrTIngredientUtil.getFluidTagInput(fluidInput2);
 
 		final RefineryRecipe recipe = new RefineryRecipe(resourceLocation, outputStack, tagInput1, tagInput2, catalyst.asVanillaIngredient(), energy);
+		recipe.modifyTimeAndEnergy(() -> 1, IEServerConfig.MACHINES.refineryConfig);
 		CraftTweakerAPI.apply(new ActionAddRecipeCustomOutput<>(this, recipe, output));
 	}
 }