+++ src/main/java/blusunrize/immersiveengineering/common/util/compat/crafttweaker/managers/SqueezerRecipeManager.java
@@ -10,6 +10,7 @@ package blusunrize.immersiveengineering.common.util.compat.crafttweaker.managers
 import blusunrize.immersiveengineering.api.crafting.IESerializableRecipe;
 import blusunrize.immersiveengineering.api.crafting.IngredientWithSize;
 import blusunrize.immersiveengineering.api.crafting.SqueezerRecipe;
+import blusunrize.immersiveengineering.common.config.IEServerConfig;
 import blusunrize.immersiveengineering.common.util.compat.crafttweaker.CrTIngredientUtil;
 import blusunrize.immersiveengineering.common.util.compat.crafttweaker.actions.AbstractActionGenericRemoveRecipe;
 import blusunrize.immersiveengineering.common.util.compat.crafttweaker.actions.ActionAddRecipeCustomOutput;
@@ -110,8 +111,8 @@ public class SqueezerRecipeManager implements IRecipeManager<SqueezerRecipe>
 		final FluidStack fluidOut = fluidOutput.getInternal();
 		final ItemStack itemOut = itemOutput.getInternal();
 
-		final SqueezerRecipe recipe = new SqueezerRecipe(
-				resourceLocation, fluidOut, IESerializableRecipe.of(itemOut), inputWithSize, energy
+		final SqueezerRecipe recipe = IEServerConfig.MACHINES.squeezerConfig.apply(
+				new SqueezerRecipe(resourceLocation, fluidOut, IESerializableRecipe.of(itemOut), inputWithSize, energy)
 		);
 
 		final String outputDescription = String.format("%s and %s", fluidOutput.getCommandString(), itemOutput.getCommandString());